feat #2 : resolve conflict
