+++ src/main/java/fastcampus/scheduling/user/repository/UserRepository.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    Optional<User> findByUserEmail(String userEmail);
-    Optional<User> findByPhoneNumber(String phoneNumber);
+	Optional<User> findByUserEmail(String userEmail);
+	Optional<User> findByPhoneNumber(String phoneNumber);
+
 }
