+++ src/main/java/fastcampus/scheduling/user/service/UserService.java
@@ -1,82 +1,57 @@
 package fastcampus.scheduling.user.service;
 
-import fastcampus.scheduling.user.dto.UserRequest;
-import fastcampus.scheduling.user.dto.UserResponse;
-import fastcampus.scheduling.user.repository.UserRepository;
+import fastcampus.scheduling._core.security.exception.AuthExceptionMessage;
+import fastcampus.scheduling.user.exception.UserExceptionMessage;
+import fastcampus.scheduling.user.exception.UserNotExistException;
 import fastcampus.scheduling.user.model.User;
-import lombok.AllArgsConstructor;
-
-import java.util.Optional;
+import fastcampus.scheduling.user.repository.UserRepository;
+import java.util.ArrayList;
+import java.util.Collection;
 import lombok.RequiredArgsConstructor;
-
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@RequiredArgsConstructor
 @Service
-public class UserService {
+@RequiredArgsConstructor
+public class UserService implements UserDetailsService {
 
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
+  private final UserRepository userRepository;
+	private final PasswordEncoder passwordEncoder;
 
-    public User findByUserId(Long id) {
-        return userRepository.findById(id)
-            .orElseThrow(() -> new IllegalStateException("유저를 찾을 수 없습니다."));
-    }
+	@Override
+  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+    User user =  userRepository.findByUserEmail(username)
+        .orElseThrow(() -> new UsernameNotFoundException(AuthExceptionMessage.MISMATCH_SIGN_IN_INFO.getMessage()));
+    Collection<SimpleGrantedAuthority> authorities = new ArrayList<>();
+    authorities.add(new SimpleGrantedAuthority(user.getPosition().name()));
+    //set name as userId(table pk)
+    return new org.springframework.security.core.userdetails.User(user.getId().toString(), user.getUserPassword(), authorities);
+  }
 
-    public User updateUser(Long id, String userPassword, String phoneNumber,
-        String profileThumbUrl) {
-        User user = userRepository.findById(id)
-            .orElseThrow(() -> new IllegalStateException("유저를 찾을 수 없습니다."));
+  public User getUserById(Long userId) {
+    return userRepository.findById(userId).orElseThrow(() -> new UserNotExistException(
+        UserExceptionMessage.USER_NOT_FOUND_EXCEPTION.getMessage()));
+  }
 
-        user.setUserPassword(userPassword);
-        user.setPhoneNumber(phoneNumber);
-        user.setProfileThumbUrl(profileThumbUrl);
+	public User findByUserId(Long id) {
+			return userRepository.findById(id)
+					.orElseThrow(() -> new IllegalStateException("유저를 찾을 수 없습니다."));
+	}
 
-        return userRepository.save(user);
-    }
-//    @Transactional
-//    public UserResponse.SignUpDTO save(UserRequest.SignUpDTO signUpDTO){
-//        if(signUpDTO == null) throw new Exception500(ErrorMessage.EMPTY_DATA_FOR_USER_SIGNUP);
-//
-//        User user = signUpDTO.toEntityWithHashPassword(passwordEncoder);
-//        User persistencUser = userRepository.save(user);
-//
-//        //access token을 만들어서 반환해줘야 할듯
-//
-//        Long id = persistencUser.getId(); //임시
-//        String userName = persistencUser.getUserName();
-//        String userInfo = id + userName;
-//
-//        String accessToken = id + userName;
-//        return UserResponse.SignUpDTO.from(accessToken);
-//    }
-//
-//    @Transactional(readOnly = true)
-//    public UserResponse.SignInDTO findByUserEmail(UserRequest.SignUpDTO signUpDTO){
-//        Optional<User> userOptional = userRepository.findByUserEmail(signUpDTO.getUserEmail());
-//
-//        return UserResponse.SignInDTO.builder()
-//            //todo
-//            .build();
-//    }
-//
-//    @Transactional(readOnly = true)
-//    public void checkPhone(UserRequest.CheckPhoneDTO checkPhoneDTO) {
-//        validatePhone(checkPhoneDTO);
-//        Optional<User> userOptional = userRepository.findByPhoneNumber(checkPhoneDTO.getPhoneNumber());
-//    }
-//
-//    private static void validatePhone(UserRequest.CheckPhoneDTO checkPhoneDTO) {
-//        //String regex = "/^[A-Za-z0-9_\\.\\-]+@[A-Za-z0-9\\-]+\\.[A-Za-z0-9\\-]+/";
-//        String phoneNumber = checkPhoneDTO.getPhoneNumber();
-//
-//        if(checkPhoneDTO == null || checkPhoneDTO.getPhoneNumber().isBlank())
-//            throw new Exception400(phoneNumber, ErrorMessage.EMPTY_DATA_FOR_USER_CHECK_PhoneNumber);
-////        if (!phoneNumber.matches(regex))
-////            throw new Exception400(phoneNumber, ErrorMessage.INVALID_PhoneNumber);
-//
-//    }
+	public User updateUser(Long id, String userPassword, String phoneNumber,
+			String profileThumbUrl) {
+			User user = userRepository.findById(id)
+					.orElseThrow(() -> new IllegalStateException("유저를 찾을 수 없습니다."));
+
+			user.setUserPassword(userPassword);
+			user.setPhoneNumber(phoneNumber);
+			user.setProfileThumbUrl(profileThumbUrl);
+
+			return userRepository.save(user);
+	}
+
 }
-
