--- src/main/java/fastcampus/scheduling/user/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package fastcampus.scheduling.user.user.repository;
-
-import fastcampus.scheduling.user.model.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-	Optional<User> findByUserEmail(String userEmail);
-	Optional<User> findByPhoneNumber(String phoneNumber);
-
-}
