--- src/main/java/fastcampus/scheduling/user/user/service/UserService.java
@@ -1,30 +1,0 @@
-package fastcampus.scheduling.user.user.service;
-
-import fastcampus.scheduling._core.security.exception.AuthExceptionMessage;
-import fastcampus.scheduling.user.model.User;
-import fastcampus.scheduling.user.user.repository.UserRepository;
-import java.util.ArrayList;
-import java.util.Collection;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService implements UserDetailsService {
-
-  private final UserRepository userRepository;
-
-  @Override
-  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-    User user =  userRepository.findByUserEmail(username)
-        .orElseThrow(() -> new UsernameNotFoundException(AuthExceptionMessage.MISMATCH_SIGN_IN_INFO.getMessage()));
-    Collection<SimpleGrantedAuthority> authorities = new ArrayList<>();
-    authorities.add(new SimpleGrantedAuthority(user.getPosition().name()));
-    //set name as userId(table pk)
-    return new org.springframework.security.core.userdetails.User(user.getId().toString(), user.getUserPassword(), authorities);
-  }
-}
